warehouse-store 1.0.0

+++ client/src/hooks/use-delete-product/use-delete-product.ts
@@ -2,39 +2,44 @@ import { useNavigate } from "react-router-dom";
 import { useAppDispatch } from "../../redux/store";
 import { closeModal, openModal, setError } from "../../redux/reducers";
 import { deleteProduct, request } from "../../utils";
-import { Dispatch, SetStateAction, useEffect, useState } from "react";
+import { Dispatch, SetStateAction } from "react";
 import { ERROR } from "../../constants";
 
 export const useDeleteProduct = (productId: string) => {
-	const [isCanDeleted, setIsCanDeleted] = useState(true);
+    const dispatch = useAppDispatch();
+    const navigate = useNavigate();
 
-	const dispatch = useAppDispatch();
-	const navigate = useNavigate();
+    return async (route: string, needRefreshPage?: boolean, setNeedRefreshPage?: Dispatch<SetStateAction<boolean>>) => {
+        try {
+            const { error, data } = await request(`/orders/check_product_in_orders`, 'POST', { productId });
 
-	useEffect(() => {
-		request(`/orders/check_product_in_orders`, 'POST', {productId}).then(({error, data}) => {
-			if (error) {
-				console.log(error);
-			} else {
-				data && setIsCanDeleted(false)
-			}
-		});
-	}, [productId])
+            if (error) {
+                console.log(error);
+                dispatch(setError(ERROR.IS_CANT_DELETE));
+                return;
+            }
 
-	return (route: string, needRefreshPage?:boolean, setNeedRefreshPage?: Dispatch<SetStateAction<boolean>>) => {
-		return isCanDeleted ?
-		dispatch(
-			openModal({
-				text: 'Удалить товар?',
-				onConfirm: () => {
-					deleteProduct(productId);
-					dispatch(closeModal());
-					navigate(route);
-					setNeedRefreshPage && setNeedRefreshPage(!needRefreshPage)
-				},
-				onCancel: () => dispatch(closeModal()),
-			}),
-		) :
-		dispatch(setError(ERROR.IS_CANT_DELETE))
-	}
-}
+            const isCanDeleted = !data;
+
+            if (isCanDeleted) {
+                dispatch(
+                    openModal({
+                        text: 'Удалить товар?',
+                        onConfirm: () => {
+                            deleteProduct(productId);
+                            dispatch(closeModal());
+                            navigate(route);
+                            setNeedRefreshPage && setNeedRefreshPage(!needRefreshPage);
+                        },
+                        onCancel: () => dispatch(closeModal()),
+                    })
+                );
+            } else {
+                dispatch(setError(ERROR.IS_CANT_DELETE));
+            }
+        } catch (err) {
+            console.error(err);
+            dispatch(setError(ERROR.IS_CANT_DELETE));
+        }
+    };
+};
